Add a test for print statements
